requirements.txt
GitPython==3.1.14
Pillow==8.1.2
py7zr==0.14.1
python-dateutil==2.8.1
PyYAML==5.4.1
rfeed==1.1.1